\name{m1}
\alias{m1}

\title{ the first-order moment of T when q(x)=exp(-mu x) }
\description{
  m1 returns the first-order moment of T when q(x)=exp(-mu x)
}
\usage{
m1(lambda, mu)
}

\arguments{
  \item{lambda}{a numeric \code{lambda} : parameter of the sequence of Xi }
  \item{mu}{ a numeric \code{mu} : parameter of q(x) }
}
\value{
  a numeric
}
\author{ CJ }


\examples{
}
\keyword{ utilities }% at least one, from doc/KEYWORDS
