\name{m2}
\alias{m2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ the second-order moment of T when q(x)=exp(-mu x) }
\description{
  m2 returns the first-order moment of T when q(x)=exp(-mu x)
}
\usage{
m2(lambda, mu)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lambda}{ a numeric \code{lambda} : parameter of the sequence of Xi }
  \item{mu}{ a numeric \code{mu} : parameter of q(x) }
}
\value{
 a numeric
}
\author{ CJ }
\examples{
}
\keyword{ utilities }% at least one, from doc/KEYWORDS
