\name{mom1}
\alias{mom1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ the first-order empirical moment of the vector X  }
\description{
  mom1 returns the first-order empirical moment of the vector X
}
\usage{
mom1(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{ a vector \code{X} on which we look for the mean }
}
\value{
  a numeric}
\author{ CJ }
\examples{
}
\keyword{ utilities }% at least one, from doc/KEYWORDS
