\name{mom2}
\alias{mom2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{the second-order empirical moment of the vector X  }
\description{
  mom2 returns the second-order empirical moment of the vector X 
}
\usage{
mom2(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{ a vector \code{X} on which we look for the second-order moment}
}
\value{
 a numeric
}
\author{ CJ }
\examples{
}
\keyword{ utilities }% at least one, from doc/KEYWORDS
